pyproject.toml
[project]
name = "trend-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]
